+++ node/utils/attachments.ts
@@ -10,7 +10,9 @@ const isParentOptionSingleChoice = ({ composition }: AssemblyOption) => {
   if (!composition) {
     return false
   }
+
   const { minQuantity, maxQuantity } = composition
+
   return minQuantity === 1 && maxQuantity === 1
 }
 
@@ -18,7 +20,9 @@ const isParentOptionToggleChoice = ({ composition }: AssemblyOption) => {
   if (!composition) {
     return false
   }
+
   const { items } = composition
+
   return all(propEq('maxQuantity', 1))(items)
 }
 
@@ -26,11 +30,15 @@ export const getItemChoiceType = (childAssemblyData?: AssemblyOption) => {
   if (!childAssemblyData) {
     return CHOICE_TYPES.MULTIPLE
   }
+
   const isSingle = isParentOptionSingleChoice(childAssemblyData!)
+
   if (isSingle) {
     return CHOICE_TYPES.SINGLE
   }
+
   const isToggle = isParentOptionToggleChoice(childAssemblyData)
+
   if (isToggle) {
     return CHOICE_TYPES.TOGGLE
   }