+++ node/utils/string.ts
@@ -2,7 +2,7 @@ export const toTitleCase = (str: string) => {
   return str
     .toLowerCase()
     .split(' ')
-    .map(word => word.charAt(0).toUpperCase() + word.slice(1))
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
     .join(' ')
 }