+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Send 1% of the production traffic to the new intsch service authough still uses the result from the I/O app. 
+
 ## [1.76.0] - 2025-02-07
 
 ### Changed